Refactoring grid title.
